tdx-tdcall: move remaining tdcall functions into tdcall.rs
