+++ tdx-tdcall/src/tdcall.rs
@@ -15,6 +15,31 @@ pub const PAGE_SIZE_4K: u64 = 0x1000;
 pub const PAGE_SIZE_2M: u64 = 0x200000;
 const TARGET_TD_UUID_NUM: usize = 4;
 
+// Guest-Side (TDCALL) interface functions leaf numbers
+const TDCALL_TDINFO: u64 = 1;
+const TDCALL_TDEXTENDRTMR: u64 = 2;
+const TDCALL_TDGETVEINFO: u64 = 3;
+const TDCALL_TDREPORT: u64 = 4;
+const TDCALL_TDACCEPTPAGE: u64 = 6;
+const TDCALL_VM_RD: u64 = 7;
+const TDCALL_VM_WR: u64 = 8;
+const TDCALL_VP_RD: u64 = 9;
+const TDCALL_VP_WR: u64 = 10;
+const TDCALL_SYS_RD: u64 = 11;
+const TDCALL_SERVTD_RD: u64 = 18;
+const TDCALL_SERVTD_WR: u64 = 20;
+const TDCALL_MEM_PAGE_ATTR_WR: u64 = 24;
+const TDCALL_VP_ENTER: u64 = 25;
+const TDCALL_VP_INVEPT: u64 = 26;
+const TDCALL_VP_INVVPID: u64 = 27;
+
+// TDCALL completion status code
+const TDCALL_STATUS_SUCCESS: u64 = 0;
+
+// leaf-specific completion status code
+pub const TDCALL_STATUS_PAGE_ALREADY_ACCEPTED: u64 = 0x00000B0A00000000;
+pub const TDCALL_STATUS_PAGE_SIZE_MISMATCH: u64 = 0xC0000B0B00000001;
+
 /// SHA384 digest value extended to RTMR
 /// Both alignment and size are 64 bytes.
 #[repr(C, align(64))]
@@ -55,6 +80,75 @@ pub struct ServtdRWResult {
     pub uuid: [u64; 4],
 }
 
+// An extended public wrapper for use of asm_td_vmcall.
+//
+// `do_sti` is a flag used to determine whether to execute `sti` instruction before `tdcall`
+pub fn td_vmcall_ex(args: &mut TdVmcallArgs, do_sti: bool) -> u64 {
+    unsafe { asm::asm_td_vmcall(args as *mut TdVmcallArgs as *mut c_void, do_sti as u64) }
+}
+
+// Wrapper for use of asm_td_call, this function takes a mutable reference of a
+// TdVmcallArgs structure to ensure the input is valid
+//
+// ## TDCALL ABI
+// Defined in TDX Module 1.0 Spec section 'TDCALL Instruction (Common)'
+//
+// ### Input Operands:
+//  * RAX - Leaf and version numbers.
+//  * Other - Used by leaf functions as input values.
+//
+// ### Output Operands:
+//  * RAX - Instruction return code.
+//  * Other - Used by leaf functions as output values.
+//
+pub fn td_call(args: &mut TdcallArgs) -> u64 {
+    unsafe { asm::asm_td_call(args as *mut TdcallArgs as *mut c_void) }
+}
+
+// Used to pass the values of input/output register when performing TDCALL
+// instruction
+#[repr(C)]
+#[derive(Default)]
+pub struct TdcallArgs {
+    pub rax: u64,
+    pub rcx: u64,
+    pub rdx: u64,
+    pub r8: u64,
+    pub r9: u64,
+    pub r10: u64,
+    pub r11: u64,
+    pub r12: u64,
+    pub r13: u64,
+}
+
+/// TDCALL instruction return error code
+///
+/// Refer to Intel TDX Module 1.0 Specifiction section 'TDCALL Instruction (Common)'
+#[derive(Debug, PartialEq)]
+pub enum TdCallError {
+    // Invalid parameters
+    TdxExitInvalidParameters,
+
+    // The operand is busy (e.g., it is locked in Exclusive mode)
+    TdxExitReasonOperandBusy(u32),
+
+    // Operand is invalid (e.g., illegal leaf number)
+    TdxExitReasonOperandInvalid(u32),
+
+    // Error code defined by individual leaf function
+    LeafSpecific(u64),
+}
+
+// TDCALL Completion Status Codes (Returned in RAX) Definition
+impl From<u64> for TdCallError {
+    fn from(val: u64) -> Self {
+        match val >> 32 {
+            0x8000_0200 => Self::TdxExitReasonOperandBusy(val as u32),
+            0xC000_0100 => Self::TdxExitReasonOperandInvalid(val as u32),
+            _ => Self::LeafSpecific(val),
+        }
+    }
+}
 
 /// Get guest TD execution environment information
 ///
